[feat] 리뷰 반영 (#3)
presentation 모듈 targetSdk 34 -> 35
MainNavController, TabController 관심사 분리(단일 책임 원칙 준수)
