+++ presentation/src/main/kotlin/com/nexters/ziine/android/presentation/navigation/TabController.kt
@@ -1,0 +1,25 @@
+package com.nexters.ziine.android.presentation.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavDestination
+import androidx.navigation.NavDestination.Companion.hasRoute
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+
+class TabController(
+    private val navController: NavHostController
+) {
+    private val currentDestination: NavDestination?
+        @Composable get() = navController.currentBackStackEntryAsState().value?.destination
+
+    val currentTab: MainTab?
+        @Composable get() = MainTab.find { tab ->
+            currentDestination?.hasRoute(tab::class) == true
+        }
+
+    @Composable
+    fun shouldShowTopBar() =
+        MainTab.contains {
+            currentDestination?.hasRoute(it::class) == true
+        }
+}
